fix: move internal utils
