+++ packages/office-ui-fabric-vue/src/Positioning.ts
@@ -1,1 +1,1 @@
-export * from './utilities/positioning'
+export * from './utils/positioning'
