+++ packages/office-ui-fabric-vue/src/utils/clickedOutside.ts
