+++ packages/office-ui-fabric-vue/src/utils/color/_rgbaOrHexString.ts
