+++ packages/office-ui-fabric-vue/src/utils/color/clamp.ts
