+++ packages/office-ui-fabric-vue/src/utils/color/getColorFromHSV.ts
