+++ packages/office-ui-fabric-vue/src/utils/color/getColorFromRGBA.ts
