+++ packages/office-ui-fabric-vue/src/utils/color/getColorFromString.ts
