+++ packages/office-ui-fabric-vue/src/utils/color/getFullColorString.ts
