+++ packages/office-ui-fabric-vue/src/utils/color/hsl2hsv.ts
