+++ packages/office-ui-fabric-vue/src/utils/color/hsl2rgb.ts
