+++ packages/office-ui-fabric-vue/src/utils/color/hsv2hex.ts
