+++ packages/office-ui-fabric-vue/src/utils/color/interfaces.ts
