+++ packages/office-ui-fabric-vue/src/utils/color/rgb2hex.ts
