+++ packages/office-ui-fabric-vue/src/utils/color/updateH.ts
