+++ packages/office-ui-fabric-vue/src/utils/color/updateRGB.ts
