+++ packages/office-ui-fabric-vue/src/utils/dragdrop/DragDropHelper.ts
