+++ packages/office-ui-fabric-vue/src/utils/dragdrop/interfaces.ts
