+++ packages/office-ui-fabric-vue/src/utils/grid/Grid.base.ts
