+++ packages/office-ui-fabric-vue/src/utils/grid/GridCell.ts
