+++ packages/office-ui-fabric-vue/src/utils/index.ts
