+++ packages/office-ui-fabric-vue/src/utils/keytips/IKeytipTransitionKey.ts
