+++ packages/office-ui-fabric-vue/src/utils/keytips/KeytipConfig.ts
