+++ packages/office-ui-fabric-vue/src/utils/keytips/KeytipManager.ts
