+++ packages/office-ui-fabric-vue/src/utils/positioning/index.ts
