+++ packages/office-ui-fabric-vue/src/utils/positioning/positioning.types.ts
