+++ packages/office-ui-fabric-vue/src/utils/responsiveMode.ts
