+++ packages/office-ui-fabric-vue/src/utils/selectableOption/SelectableOption.ts
@@ -1,4 +1,4 @@
-import { ISelectableOption } from '../../utilities/selectableOption/SelectableOption.types'
+import { ISelectableOption } from '../selectableOption/SelectableOption.types'
 
 export function getAllSelectedOptions (options: ISelectableOption[], selectedIndices: number[]): ISelectableOption[] {
   const selectedOptions: ISelectableOption[] = []
