+++ packages/office-ui-fabric-vue/src/utils/selection/Selection.ts
