+++ packages/office-ui-fabric-vue/src/utils/selection/index.ts
