+++ packages/office-ui-fabric-vue/src/utils/selection/interfaces.ts
