+++ packages/office-ui-fabric-vue/src/utils/semanticColors.ts
